blackbox: add ring-buffer flash logging mode (USE_BLACKBOX_RING_LOG)
Adds a circular-buffer onboard-flash logging mode for blackbox, gated by a new
CLI parameter `blackbox_flash_mode = LINEAR | RING` (default LINEAR — existing
users see no change). Linear-mode code paths are bit-for-bit untouched.

Layout (ring mode only):

  [ data ring ........................................ ][ active hdr buffer ]
    0                                                    partition - 16 KB

Each log writes its data into the ring, then on clean close writes a small
sector-aligned trailer record followed by the full header text at the END of
its data:

    ... [data of log N][trailer][header text][data of log N+1][trailer][header text] ...

Eviction is implicit: when ring writes overrun an old log's trailer, that log
disappears. No metadata bookkeeping, no INVALIDATE records.

Erase pool (NEW — required for correctness on NOR flash)
--------------------------------------------------------
NOR flash writes are AND-only — programming a sector that hasn't been erased
corrupts data. To support multi-lap ring writes, the module maintains an erase
frontier `eraseHead` running ahead of the writer. Erases happen ASYNCHRONOUSLY
in the chip's background (CPU returns immediately from flashEraseSector; the
chip is busy ~30-50 ms while writes pile up in the RAM buffer).

  - At log start: pre-erase POOL_TARGET_SECTORS=4 sectors ahead synchronously
  - During flight: eraseTick() in the data write hot path checks erase
    completion and triggers the next async erase. Never blocks the CPU.
  - At log close / recovery: ensureErasedSpace() sync-erases additional sectors
    as needed to fit the trailer + header (header may span 2-3 sectors)

Flashfs RAM buffer enlarged to 16 KB (was 128 B) when ring log is compiled in,
sized to absorb one worst-case erase stall at the maximum allowed sample rate.
This requires `bufferHead`/`bufferTail` to widen from uint8_t to uint16_t.

Sample rate clamp: blackboxValidateConfig() forces sample_rate ≥ 1/4 in ring
mode (~200 KB/s ≤ ~133 KB/s NOR erase rate doesn't quite hold but it's close).
1/8 is the recommended setting for gap-free logs.

Drop discipline: data write hot path drops bytes when the writer catches the
erase frontier or the RAM buffer is full. Lost frames appear as gaps in the
log; blackbox-log-viewer handles them.

Power-loss safety
-----------------
Active log's header is buffered in a 4-sector reserved area (16 KB) at the top
of the partition. At clean log close we copy header → end-of-data, mark
committed, and erase the buffer. On boot:

  - Buffer empty             → no recovery
  - Buffer valid, uncommitted → power loss; scan the entire ring for the
                                erased gap left by the erase pool, then write
                                trailer + header there. If header is bigger
                                than the gap, sync-erase additional sectors.
                                Recovered log = "everything in the ring except
                                the erased gap" — gives the user the most
                                recent ring-worth of frames regardless of
                                how many laps the flight did.
  - Buffer valid, committed   → trailer write succeeded but the buffer erase
                                was interrupted; just erase

MSC presentation
----------------
Each ring-mode log appears as a virtual emfat file whose contents are
header || data. blackbox-log-viewer needs no changes — it sees the same
legacy layout it always has.

Build flag
----------
USE_BLACKBOX_RING_LOG enabled by default on STM32F7 / STM32H7 only (Cortex-M7
with sufficient SRAM for the enlarged 16 KB flashfs buffer). F4 and smaller
targets can opt in via cloud build option. When disabled, all ring-mode call
sites compile to inline-stub no-ops with no link dependency on flashfs_log.c.

Mode safety
-----------
Switching between linear and ring requires an erase. Both writer paths refuse
to operate on a chip whose on-flash format doesn't match the configured mode.

NEEDS HARD-REALTIME EXPERT REVIEW:
  - eraseTick() in flashfsLogWriteDataByte hot path: claims to never block
    (only flashIsReady() register read + flashEraseSector() command-submit).
    Confirm against actual driver behavior on each supported flash chip.
  - flashfsLogBeginLog erases ~5 sectors synchronously (~150-300 ms total
    across header buffer + erase pool). Sanity-check the existing
    PREPARE_LOG_FILE / SEND_HEADER state pacing tolerates this.
  - flashfsLogEndLog writes trailer + full header sync at disarm and may
    sync-erase additional sectors. Confirm against
    BLACKBOX_SHUTDOWN_TIMEOUT_MILLIS (200 ms).
  - Data ring wrap (flashfsSeekAbs(0)) blocks on flashfsFlushSync until
    pending writes drain. Happens ~once per ring loop (~14 MB).
  - Sample rate clamp at 1/4 is conservative; tune against measured
    sustained rates on real NOR / NAND chips.

+++ src/main/io/flashfs.c
@@ -71,8 +71,10 @@ static DMA_DATA_ZERO_INIT uint8_t flashWriteBuffer[FLASHFS_WRITE_BUFFER_SIZE];
  * The tail is advanced once a write is complete up to the location behind head. The tail is advanced
  * by a callback from the FLASH write routine. This prevents data being overwritten whilst a write is in progress.
  */
-static uint8_t bufferHead = 0;
-static volatile uint8_t bufferTail = 0;
+// Type-wide enough to index FLASHFS_WRITE_BUFFER_SIZE; bumped from uint8_t when the
+// buffer was 128 B to uint16_t now that USE_BLACKBOX_RING_LOG can grow it to 16 KB.
+static uint16_t bufferHead = 0;
+static volatile uint16_t bufferTail = 0;
 
 /* Track if there is new data to write. Until the contents of the buffer have been completely
  * written flashfsFlushAsync() will be repeatedly called. The tail pointer is only updated
